requirements.txt
streamlit
langchain-community
langchain-core
faiss-cpu
huggingface-hub
sentence-transformers
transformers
python-dotenv
PyPDF2
google-genai
torch
